backend/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-cors==6.0.2
psycopg-binary==3.3.2
python-dotenv==1.2.1
PyJWT==2.10.1
celery==5.6.1
redis==7.1.0
requests==2.32.5
reportlab==4.4.7
pandas==2.3.3
scikit-learn==1.8.0